app/requirements.txt
asgiref==3.7.2
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
Django==4.2.3
django-debug-toolbar==4.1.0
fonttools==4.40.0
html5lib==1.1
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
pydyf==0.7.0
pyphen==0.14.0
python-dateutil==2.8.2
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.7.1
weasyprint==59.0
webencodings==0.5.1
zopfli==0.2.2
